+++ double-conversion/double-conversion.h
@@ -533,7 +533,7 @@ class StringToDoubleConverter {
                       bool read_as_double,
                       int* processed_characters_count) const;
 
-  double StringToIeeeOnAsciiBuffer(char* buffer,
+  float StringToIeeeOnAsciiBuffer(char* buffer,
                                    int length,
                                    bool read_as_double,
                                    int* processed_characters_count) const;
